[artf7275] : Backup jobs are failing due to missing logs/temp folder
Now ensures that /logs/temp exists when a progress log file name is generated; /logs/temp is also skipped by the log rotate / cleanup job.

* LogRotateJob.groovy (pruneLog)
keeps /logs/temp during log cleanup job

* BackgroundJobUtil.groovy (prepareProgresssLogFile)
now creates /logs/temp if it doesn't exist

git-svn-id: https://ctf.open.collab.net/svn/repos/svnedge/trunk@3210 03e8f217-bfc6-4b7c-bcb7-0738c91e2c5f

+++ grails-app/jobs/com/collabnet/svnedge/admin/LogRotateJob.groovy
@@ -65,6 +65,10 @@ class LogRotateJob {
         for (int i = 0; i < entries.length; i++) {
             String entry = entries[i]
             def entryObj = new File(entry)
+            // do not remove the /logs/temp folder
+            if (entryObj.isDirectory() && entryObj.name == "temp") {
+                continue
+            }
             long lastModified = entryObj.lastModified()
             if (lastModified < pruningTimestamp) {
                 if (entryObj.delete() == false) {
